+++ src/components/Day/TodayNote/index.tsx
@@ -1,32 +1,12 @@
-import { useState } from 'react';
+import useDebouncing from 'src/hooks/useDebouncing';
+import { theme } from 'src/styles/theme';
 import styled from 'styled-components';
 
 function TodayNote() {
-  const [todayNote, setTodayNote] = useState<string>();
-  const [timer, setTimer] = useState<NodeJS.Timeout | number>(0);
-
-  const handleChange = (e: React.ChangeEvent<HTMLTextAreaElement>) => {
-    setTodayNote(e.target.value);
-    if (timer) {
-      clearTimeout(timer);
-    }
-    const newTimer = setTimeout(async () => {
-      try {
-        //서버와 api 통신
-        console.log('저장');
-      } catch (e) {
-        console.error('error', e);
-      }
-    }, 2500);
-    setTimer(newTimer);
-  };
+  const { ...debouncingInfo } = useDebouncing();
   return (
     <StyledTodayNote.Root>
-      <StyledTodayNote.Textarea
-        placeholder="오늘 하루 노트를 작성해보세요."
-        onChange={handleChange}
-        value={todayNote}
-      />
+      <StyledTodayNote.Textarea placeholder="오늘 하루 노트를 작성해보세요." {...debouncingInfo} />
     </StyledTodayNote.Root>
   );
 }
@@ -39,11 +19,11 @@ const StyledTodayNote = {
     height: 11rem;
     margin-top: 1.2rem;
     padding: 1.6rem 2rem;
-    background-color: #f8f9fb;
+    background-color: ${theme.colors.scroll_grey};
     border-radius: 0.2rem;
   `,
   Textarea: styled.textarea`
-    background-color: #f8f9fb;
+    background-color: ${theme.colors.scroll_grey};
     border: none;
     padding: 0;
     width: 100%;
@@ -51,7 +31,11 @@ const StyledTodayNote = {
     font-size: 1.2rem;
     font-weight: 400;
     &::placeholder {
-      color: #d8dbde;
+      color: ${theme.colors.letter_grey};
+    }
+    &:focus {
+      color: ${theme.colors.letter_black};
+      outline: 0;
     }
   `,
 };
